requirements.txt
beautifulsoup4==4.6.0
chardet==3.0.4
click==7.1.2
discord.py==1.6.0
konlpy==0.5.2
twitter==1.18.0
requests==2.25.1
smtp==0.1.0
mime==0.1.0